refactor simulated datasets
