--- docs/api/simulated_datasets.md
@@ -1,8 +1,0 @@
-# Simulated Data
-
-## <span class="doc-module-prefix">module</span> glide.core.simulated_datasets
-
-::: glide.core.simulated_datasets
-    options:
-      heading_level: 3
-      show_root_heading: false
