+++ docs/api/simulators.md
@@ -1,0 +1,8 @@
+# Simulated Data
+
+## <span class="doc-module-prefix">module</span> glide.simulators
+
+::: glide.simulators
+    options:
+      heading_level: 3
+      show_root_heading: false
